+++ src/datumaro/components/abstracts/merger.py
@@ -5,6 +5,7 @@
 from abc import ABC, abstractmethod
 from typing import Dict, Optional, Sequence, Type
 
+from datumaro.components.annotation import Annotation
 from datumaro.components.dataset_base import IDataset
 from datumaro.components.dataset_item_storage import (
     DatasetItemStorage,
@@ -12,10 +13,16 @@ from datumaro.components.dataset_item_storage import (
 )
 from datumaro.components.media import MediaElement
 
-__all__ = ["IMerger"]
+__all__ = ["IMatcherContext", "IMergerContext"]
 
 
-class IMerger(ABC):
+class IMatcherContext(ABC):
+    @abstractmethod
+    def get_any_label_name(self, ann: Annotation, label_id: int) -> str:
+        raise NotImplementedError
+
+
+class IMergerContext(IMatcherContext):
     @abstractmethod
     def merge_infos(self, sources: Sequence[IDataset]) -> Dict:
         raise NotImplementedError